api/requirements.txt
django==5.0.*
djangorestframework==3.15.*
celery==5.4.*
redis==5.0.*
psycopg2-binary==2.9.*
pandas==2.2.*
openpyxl==3.1.*
gunicorn==22.0.*
python-dotenv==1.0.*